# backup/20241230_000244/api-docs.yaml
openapi: 3.0.0
info:
  title: Heymarket API
  version: 1.0.0
  description: |
    API for interacting with Heymarket messaging service.
    Supports single messages, templates, and batch operations.

    ## Template Variables
    When using templates, variables are specified in {{variable}} format. Variables must match exactly between the template and the provided variables object.

    Example:
    ```
    Template: "Hello {{name}}, your points: {{points}}"
    Variables: { "name": "John", "points": "100" }
    ```

    Note: Dot notation (e.g. {{customer.name}}) requires the variable name to match exactly:
    ```
    Template: "Hello {{customer.name}}"
    Variables: { "customer.name": "John" }  # Must match template variable exactly
    ```
servers:
  - url: https://heymarket-api-dpbubhdceqb3fvge.canadaeast-01.azurewebsites.net
    description: Production server
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
paths:
...
  /api/messages/batch:
    post:
      summary: Create a batch message operation
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - template
                - recipients
              properties:
                batchId:
                  type: string
                  description: Optional unique identifier for idempotency
                template:
                  type: object
                  oneOf:
                    - type: object
                      required:
                        - id
                      properties:
                        id:
                          type: string
                          description: ID of an existing template
                    - type: object
                      required:
                        - text
                      properties:
                        text:
                          type: string
                          description: |
                            Custom message text with variables in {{variable}} format.
                            Variables in the template must match exactly with the variables provided.
                            Example: "Hello {{name}}" requires variables: {"name": "John"}
                            Example with dot notation: "Hello {{customer.name}}" requires variables: {"customer.name": "John"}
                          example: "Hello {{name}}, your points: {{points}}"
                        attachments:
                          type: array
                          description: Optional array of media URLs
                          items:
                            type: string
                        isPrivate:
                          type: boolean
                          description: Whether messages are private
                        author:
                          type: string
                          description: Author for the messages
                recipients:
                  type: array
                  items:
                    type: object
                    required:
                      - phoneNumber
                      - variables
                    properties:
                      phoneNumber:
                        type: string
                      variables:
                        type: object
                        description: |
                          Variables must match template variables exactly.
                          For dot notation (e.g. {{customer.name}}), use "customer.name" as the key.
                        additionalProperties:
                          type: string
                        example:
                          name: "John"
                          points: "100"
                      overrides:
                        type: object
                        properties:
                          isPrivate:
                            type: boolean
                          author:
                            type: string
                options:
                  type: object
                  properties:
                    scheduleTime:
                      type: string
                      format: date-time
                    priority:
                      type: string
                      enum: [high, normal, low]
                    retryStrategy:
                      type: object
                      properties:
                        maxAttempts:
                          type: integer
                        backoffMinutes:
                          type: integer
      responses:
        '201':
          description: Batch created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      batchId:
                        type: string
                      status:
                        type: string
                        enum: [pending, processing, completed, failed]
                      progress:
                        type: object
                        properties:
                          total:
                            type: integer
                          pending:
                            type: integer
                          processing:
                            type: integer
                          completed:
                            type: integer
                          failed:
                            type: integer
  /api/batch/history:
    get:
      summary: Get batch request history and active batches
      security:
        - ApiKeyAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
          description: Number of history entries to return
      responses:
        '200':
          description: History retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      history:
                        type: array
                        items:
                          type: object
                          properties:
                            timestamp:
                              type: string
                              format: date-time
                            method:
                              type: string
                            path:
                              type: string
                            body:
                              type: object
                            query:
                              type: object
                      activeBatches:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                            status:
                              type: string
                            created:
                              type: string
                              format: date-time
                            template:
                              type: object
                              properties:
                                id:
                                  type: string
                                text:
                                  type: string
                            progress:
                              type: object
                              properties:
                                total:
                                  type: integer
                                pending:
                                  type: integer
                                processing:
                                  type: integer
                                completed:
                                  type: integer
                                failed:
                                  type: integer
                            metrics:
                              type: object
                              properties:
                                messages_per_second:
                                  type: number
                                success_rate:
                                  type: number
                                credits_used:
                                  type: number
  /api/batch/{batchId}:
    get:
      summary: Get batch operation status
      security:
        - ApiKeyAuth: []
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Batch status retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      status:
                        type: string
                      progress:
                        type: object
                      timing:
                        type: object
                      errors:
                        type: object
                      metrics:
                        type: object
                        properties:
                          messages_per_second:
                            type: number
                          success_rate:
                            type: number
                          credits_used:
                            type: number
  /api/batch/{batchId}/results:
    get:
      summary: Get results of a batch operation
      security:
        - ApiKeyAuth: []
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Batch results retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      batchId:
                        type: string
                      results:
                        type: array
                        items:
                          type: object
                          properties:
                            phoneNumber:
                              type: string
                            status:
                              type: string
                              enum: [success, skipped, failed]
                            messageId:
                              type: string
                            timestamp:
                              type: string
                              format: date-time
                            attempts:
                              type: integer
  /api/batch/{batchId}/analytics:
    get:
      summary: Get analytics for a batch operation
      security:
        - ApiKeyAuth: []
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Batch analytics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      messages_per_second:
                        type: number
                      success_rate:
                        type: number
                      completionRate:
                        type: number
                      errorBreakdown:
                        type: object
                      timing:
                        type: object
  /api/batch/{batchId}/errors:
    get:
      summary: Get errors from a batch operation
      security:
        - ApiKeyAuth: []
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Batch errors retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      batchId:
                        type: string
                      total:
                        type: integer
                      errors:
                        type: array
                        items:
                          type: object
                          properties:
                            phoneNumber:
                              type: string
                            error:
                              type: string
                            errorCategory:
                              type: string
                              enum: [rate_limit, invalid_request, network_error, timeout]
                            timestamp:
                              type: string
                              format: date-time
                            attempts:
                              type: integer